LPS-55718 Moves EditRoleAssignmentsAction, no logic changes
