--- portal-impl/src/com/liferay/portlet/rolesadmin/action/EditRoleAssignmentsAction.java
@@ -1,158 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.rolesadmin.action;
-
-import com.liferay.portal.NoSuchRoleException;
-import com.liferay.portal.RoleAssignmentException;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.model.Role;
-import com.liferay.portal.model.RoleConstants;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.security.membershippolicy.MembershipPolicyException;
-import com.liferay.portal.service.GroupServiceUtil;
-import com.liferay.portal.service.RoleLocalServiceUtil;
-import com.liferay.portal.service.UserServiceUtil;
-import com.liferay.portal.struts.PortletAction;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class EditRoleAssignmentsAction extends PortletAction {
-
-	@Override
-	public void processAction(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		try {
-			if (cmd.equals("role_groups")) {
-				updateRoleGroups(actionRequest);
-			}
-			else if (cmd.equals("role_users")) {
-				updateRoleUsers(actionRequest);
-			}
-
-			if (Validator.isNotNull(cmd)) {
-				String redirect = ParamUtil.getString(
-					actionRequest, "assignmentsRedirect");
-
-				sendRedirect(actionRequest, actionResponse, redirect);
-			}
-		}
-		catch (Exception e) {
-			if (e instanceof MembershipPolicyException) {
-				SessionErrors.add(actionRequest, e.getClass(), e);
-			}
-			else if (e instanceof NoSuchRoleException ||
-					 e instanceof PrincipalException ||
-					 e instanceof RoleAssignmentException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-
-				setForward(actionRequest, "portlet.roles_admin.error");
-			}
-			else {
-				throw e;
-			}
-		}
-	}
-
-	@Override
-	public ActionForward render(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		try {
-			ActionUtil.getRole(renderRequest);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchRoleException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return actionMapping.findForward("portlet.roles_admin.error");
-			}
-			else {
-				throw e;
-			}
-		}
-
-		return actionMapping.findForward(
-			getForward(
-				renderRequest, "portlet.roles_admin.edit_role_assignments"));
-	}
-
-	protected void updateRoleGroups(ActionRequest actionRequest)
-		throws Exception {
-
-		long roleId = ParamUtil.getLong(actionRequest, "roleId");
-
-		long[] addGroupIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "addGroupIds"), 0L);
-		long[] removeGroupIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "removeGroupIds"), 0L);
-
-		Role role = RoleLocalServiceUtil.getRole(roleId);
-
-		if (role.getName().equals(RoleConstants.OWNER)) {
-			throw new RoleAssignmentException(role.getName());
-		}
-
-		GroupServiceUtil.addRoleGroups(roleId, addGroupIds);
-		GroupServiceUtil.unsetRoleGroups(roleId, removeGroupIds);
-	}
-
-	protected void updateRoleUsers(ActionRequest actionRequest)
-		throws Exception {
-
-		long roleId = ParamUtil.getLong(actionRequest, "roleId");
-
-		long[] addUserIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "addUserIds"), 0L);
-		long[] removeUserIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "removeUserIds"), 0L);
-
-		Role role = RoleLocalServiceUtil.getRole(roleId);
-
-		if (role.getName().equals(RoleConstants.OWNER)) {
-			throw new RoleAssignmentException(role.getName());
-		}
-
-		UserServiceUtil.addRoleUsers(roleId, addUserIds);
-		UserServiceUtil.unsetRoleUsers(roleId, removeUserIds);
-	}
-
-}
